To improve Sabre's performance, several implementations have been updated. I’ve added comments to the code and each comment number corresponds to the explanation numbers provided below. Please let me know if any clarifications are needed. Below is the code explanation.

Checklist:
- [ ] Reviewers confirm new code works as expected.
- [x] Tests are passing.
- [x] Coverage does not decrease.
- [x] Documentation is updated.

After the code review:
- [x] Delete comments.
- [x] Update test functions.
- [ ] Optimize `_swap_candidates(), _check_execution(), _shortest_path_routing()` (remove `get_physical_qubits(block_num)`).
- [ ] Optimize `ShortestPaths`.