@csookim 2024-08-15
# SABRE + CircuitMap Optimization

## Previous Implementation

- The mapping scheme could be clearer.
- Using `list.index()` in logical-to-physical and circuit-to-physical mapping slowed down the routing.
- Frequently using the function that finds a gate's target qubits by block number was inefficient.

## New Implementation

The code has been updated in three key areas:

### 1. Qubit Mapping

- SABRE operates using `connectivity`, `initial_layout`, and `circuit`.
- `initial_layout` is a dictionary where `"q{physical_qubit_number}": {logical_qubit_number}` pairs define the initial mapping.
- `physical_qubit_number` corresponds to the node number in `connectivity`.
- The `initial_layout` and `connectivity` can have arbitrary `physical_qubit_numbers`, when using restricted connectivity.

#### 1-1. Layout Relabel

- To simplify routing, `initial_layout` and `connectivity` are relabeled before routing begins.
- `physical_qubit_number` is reassigned to a range from `0` to `len(connectivity.nodes) - 1`.
- For example, the layout `{"q4":1, "q6":2, "q3":0}` might be relabeled to `{"q0":1, "q1":2, "q2":0}`.
- This relabeling map is saved, and the final layout is reverted to the original labeling when routing is complete.
- All operations in `Sabre` are performed using the updated `physical_qubit_number`.

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L731-L739

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L702-L707

#### 1-2 Physical-Logical Mapping

- Due to the layout labeling, there are two sets of numbers ranging from `0` to `len(connectivity.nodes()) - 1` in `initial_layout`.
- Physical to logical mapping (`p2l`) and logical to physical mapping (`l2p`) are implemented using two arrays: `self._p2l` and `self._l2p`.
- For example, with a mapping of `{"q0":1, "q1":2, "q2":0}`, `self._p2l` would be `[1, 2, 0]` and `self._l2p` would be `[2, 0, 1]`.
- Routing is performed using these p2l and l2p conversions.

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L183-L192

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L280-L284
 
### 2. Temporary CircuitMap

- Creating a `CircuitMap` includes unnecessary operations such as block decomposition, circuit object creation, and gate copying, which are not needed for calculating the cost of a candidate swap.
- Enable temporary `CircuitMap` using a `self._temporary` flag.
- For temporary `CircuitMap` instances used only to calculate swap costs, return immediately from `init` after initializing `self._l2p` and `self._p2l`.
- In these temporary `CircuitMap` instances, skip gate additions or removals in the `update()` method.

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L194-L196

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L843-L849

### 3. Blocks in the `front_layer`

- SABRE calculates the cost of a swap by summing the distances of successive gates.
- To calculate the distance of a successive gate, the target qubit numbers of the gate need to be given.
- Previously, only gate (block) numbers were stored in the `front_layer`.
- To find the target qubits, the code had to iterate through the list of gates using `search_by_index(block)`, which increased runtime.
- The new approach stores both the gate number and the target qubits when creating `self._dag`.

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L990-L995

- For example,

```py
self._dag = [(0, {'qubits': (0, 9), 'layer': 0}), (1, {'qubits': (5, 9), 'layer': 1}), (2, {'qubits': (2, 8), 'layer': 0})]
```

- The function `_get_dag_layer(n, qubits=True)` returns a list of target qubits for the gates in layer `n`.
- This allows retrieval of target qubits in $O(1)$ time.

https://github.com/qiboteam/qibo/blob/0111a6e49bc27b8e5b8102f5eb6615a093ade35c/src/qibo/transpiler/router.py#L807-L815

### 4. Additional Improvements

- Use shallow copy when copying integer lists.

## Evaluation

- Test: 100 random 50-CZ gates on 20 qubits.
- Qibo SABRE is 6x faster than before.
- The reason for the reduction in the number of SWAP gates is unknown.

```py
                     Mean   Std
Qiskit Sabre Time    0.0439 0.0075
Qibo Sabre Time      0.0625 0.0179
Qibo Sabre Old Time  0.4252 0.106

Qiskit Sabre Swaps   189.97 21.7295
Qibo Sabre Swaps     196.54 21.8712
Qibo Sabre Old Swaps 279.96 35.739
```

<img width="891" alt="Screenshot 2024-08-14 at 20 49 23" src="https://github.com/user-attachments/assets/13be11e9-7913-4989-9170-cd2d59ca7492">