@csookim 2024-08-28
@BrunoLiegiBastonLiegi Thank you for your review. However, if we consider using a custom string for the physical qubit it would result in wide code modifications, such as updates to test code and `placer`.

As far as I've heard, qibolab isn't ready for string-based qubit names. So, I suggest we complete this optimization PR with the `"q" + str(i)` physical qubit names. Then, we can open a new PR focused on establishing a unified qubit naming scheme for both qibo and qibolab. I'd like to hear your thoughts on this approach.