@BrunoLiegiBastonLiegi 2024-08-29
Thanks @csookim for the insight about the `wire_names`, indeed as this PR is already adding various optimizations to the `CircuitMap` object, I am ok with merging this with the support for the default naming scheme only (`q0, q1, ...`), as long as we open an issue to remind us about it. So, in the end, up to you, if you feel this is something straightforward we could address it here directly, fine, otherwise we leave it for a future PR. In any case the support to custom qubit names is probably not a widely used feature anyway...