@Simone-Bordoni 2024-08-30
> Custom qubit names can be used in Sabre. But the node names in the `connectivity graph` and the physical qubit names in the `layout` must be identical. Layout relabeling and recovery are done within `CircuitMap`.
> 
> ### `CircuitMap`
> * Receive the layout `{'A': 2, 'B': 0, 'C': 1}`.
> * Assign physical qubit numbers to the keys of the layout (`'A' -> 0`, `'B' -> 1`, `'C' -> 2`).
> * Relabel the connectivity graph based on this mapping.
> * Save the qubit names (keys) in `wire_names`.
> * Logical qubit numbers are the values of the layout.
> * Route using physical and logical qubit numbers.
> * Restore the original connectivity graph using the inverse of the mapping (`0 -> 'A'`, `1 -> 'B'`, `2 -> 'C'`).
> 
> Regarding the test files, some `placers` currently use node names that do not match the physical qubit names.
> 
> https://github.com/qiboteam/qibo/blob/21732ad59d31740e8c0641921abe98e49a1ea8c1/src/qibo/transpiler/placer.py#L203-L205
> 
> I believe some parts of the code in `Placer` need to be updated, and the test files should be updated after resolving issue #1429.

I agree on relabelling the graph nodes with the physical qubit names. However, in principle, the two objects should already have the same names. It is better to raise at least a warning when the names do not match, something like "Connectivity graph nodes names and physical qubit names dont match. Renaming connectivity nodes after physical qubits."